Temporary fix for GPU process crash on windows
The GPU process crashes on windows for some reason when we
compile with MSVC but not when we compile with MinGW (maybe stdlib-related?).  Happens
in libcef in a call to CreateWindowsOnThread, which is
being used by chrome for something directx-related.

The debugger says CreateWindowEx returns NULL but that GetLastError() is ERROR_SUCCESS,
suggesting something wrong with the WndProc.

https://chromium.googlesource.com/chromium/src/+/refs/heads/main/ui/gl/child_window_win.cc#108
https://github.com/chromiumembedded/cef/issues/3765